frontend/requirements.txt
pandas
plotly
pillow
streamlit
requests
st-pages

